--test_list_tests: do not show unknown type value

+++ test/simple.c
@@ -10,7 +10,7 @@ TEST_FIXTURE_SETUP(simple)
 TEST_FIXTURE_TEAREDOWN(simple)
 {
 }
-TEST_PARAMETERIZED_DEFINE(simple, int, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+TEST_PARAMETERIZED_DEFINE(simple, size_t, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
 
 static int s_sum;
